odoogpt_cu/requirements.txt
openai
python-dotenv
psycopg2-binary
phonenumbers
flanker
markdown2
markupsafe
PyPDF2
python-docx
openpyxl
python-dateutil